+++ sdk/data/aztables/transactional_batch.go
@@ -119,7 +119,7 @@ func (t *Client) submitTransactionInternal(ctx context.Context, transactionActio
 	if err != nil {
 		return TransactionResponse{}, err
 	}
-	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(t.client.Con.Endpoint(), "$batch"))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(t.con.Endpoint(), "$batch"))
 	if err != nil {
 		return TransactionResponse{}, err
 	}
@@ -154,7 +154,7 @@ func (t *Client) submitTransactionInternal(ctx context.Context, transactionActio
 		return TransactionResponse{}, err
 	}
 
-	resp, err := t.client.Con.Pipeline().Do(req)
+	resp, err := t.con.Pipeline().Do(req)
 	if err != nil {
 		return TransactionResponse{}, err
 	}
@@ -308,12 +308,27 @@ func (t *Client) generateEntitySubset(transactionAction *TransactionAction, writ
 
 	switch transactionAction.ActionType {
 	case Delete:
-		req, err = t.client.DeleteEntityCreateRequest(ctx, t.name, entity[partitionKey].(string), entity[rowKey].(string), string(*transactionAction.IfMatch), &generated.TableDeleteEntityOptions{}, qo)
+		req, err = t.client.DeleteEntityCreateRequest(
+			ctx,
+			generated.Enum1Three0,
+			t.name,
+			entity[partitionKey].(string),
+			entity[rowKey].(string),
+			string(*transactionAction.IfMatch),
+			&generated.TableDeleteEntityOptions{},
+			qo,
+		)
 		if err != nil {
 			return err
 		}
 	case Add:
-		req, err = t.client.InsertEntityCreateRequest(ctx, t.name, &generated.TableInsertEntityOptions{TableEntityProperties: entity, ResponsePreference: generated.ResponseFormatReturnNoContent.ToPtr()}, qo)
+		req, err = t.client.InsertEntityCreateRequest(
+			ctx,
+			generated.Enum1Three0,
+			t.name,
+			&generated.TableInsertEntityOptions{TableEntityProperties: entity, ResponsePreference: generated.ResponseFormatReturnNoContent.ToPtr()},
+			qo,
+		)
 		if err != nil {
 			return err
 		}
@@ -324,17 +339,32 @@ func (t *Client) generateEntitySubset(transactionAction *TransactionAction, writ
 		if transactionAction.IfMatch != nil {
 			opts.IfMatch = to.StringPtr(string(*transactionAction.IfMatch))
 		}
-		req, err = t.client.MergeEntityCreateRequest(ctx, t.name, entity[partitionKey].(string), entity[rowKey].(string), opts, qo)
+		req, err = t.client.MergeEntityCreateRequest(
+			ctx,
+			generated.Enum1Three0,
+			t.name,
+			entity[partitionKey].(string),
+			entity[rowKey].(string),
+			opts,
+			qo,
+		)
 		if err != nil {
 			return err
 		}
-		if isCosmosEndpoint(t.client.Con.Endpoint()) {
+		if isCosmosEndpoint(t.con.Endpoint()) {
 			transformPatchToCosmosPost(req)
 		}
 	case UpdateReplace:
 		fallthrough
 	case InsertReplace:
-		req, err = t.client.UpdateEntityCreateRequest(ctx, t.name, entity[partitionKey].(string), entity[rowKey].(string), &generated.TableUpdateEntityOptions{TableEntityProperties: entity, IfMatch: to.StringPtr(string(*transactionAction.IfMatch))}, qo)
+		req, err = t.client.UpdateEntityCreateRequest(
+			ctx,
+			generated.Enum1Three0,
+			t.name,
+			entity[partitionKey].(string), entity[rowKey].(string),
+			&generated.TableUpdateEntityOptions{TableEntityProperties: entity, IfMatch: to.StringPtr(string(*transactionAction.IfMatch))},
+			qo,
+		)
 		if err != nil {
 			return err
 		}